+++ src/components/Panel.js
@@ -1,6 +1,7 @@
 import React from "react";
 
 import classnames from "../classnames";
+import Loader from "./Loader";
 
 const Panel = ({ className, children }) => (
   <div className={classnames("chq-pan", className)}>{children}</div>
@@ -17,6 +18,12 @@ const PanelBody = ({ className, children }) => (
   <div className={classnames("chq-pan--bd", className)}>{children}</div>
 );
 
+const PanelLoaderBody = ({ className, children, loading }) => (
+  <PanelBody className={className}>
+    <Loader loading={loading}>{children}</Loader>
+  </PanelBody>
+);
+
 const PanelFooter = ({ className, children }) => (
   <div className={classnames("chq-pan--ft", className)}>{children}</div>
 );
@@ -24,6 +31,7 @@ const PanelFooter = ({ className, children }) => (
 Object.assign(Panel, {
   Heading: PanelHeading,
   Body: PanelBody,
+  LoaderBody: PanelLoaderBody,
   Footer: PanelFooter
 });